A,B,C repair calc algorithm, transfer to double

+++ determinant.h
@@ -6,12 +6,12 @@
 
 namespace math {
 
-using M3x3 = std::array<float, 9>;
-using M2x2 = std::array<float, 4>;
+using M3x3 = std::array<double, 9>;
+using M2x2 = std::array<double, 4>;
 
 struct LineIntersectionPoint {
-  float x;
-  float y;
+  double x;
+  double y;
   unsigned int L1;
   unsigned int L2;
   unsigned int L3;
@@ -22,14 +22,14 @@ struct LineIntersectionPoint {
  * @param m - matrix 3x3
  * @return
  */
-float Determinant3x3(const M3x3 &m);
+double Determinant3x3(const M3x3 &m);
 
 /**
  * @brief Determinant2x2
  * @param m
  * @return
  */
-float Determinant2x2(const M2x2 &m);
+double Determinant2x2(const M2x2 &m);
 
 /**
  * @brief CalcIntersectionPoint
